requirements.txt
confluent-kafka
pyshark
PyYAML

